Add config module with configure function
